// appointza-ui-canvas/src/models/users.model.ts
export class Users {
  id: number = 0;
  name: string = '';
  email: string = '';
  mobile: string = '';
  mobilecountrycode: string = '';
  designation: string = '';
  otp: string = '';
  otpexpirationtime: Date = new Date();
  organisationid: number = 0;
  locationid: number = 0;
  version: number = 0;
  createdby: number = 0;
  createdon: Date = new Date();
  modifiedby: number = 0;
  modifiedon: Date = new Date();
  attributes: Users.AttributesData = new Users.AttributesData();
  isactive: boolean = false;
  issuspended: boolean = false;
  parentid: number = 0;
  isfactory: boolean = false;
  notes: string = '';
  profileimage: number = 0;
}

export namespace Users {
  export class AttributesData {
    permission: UsersPermissionData = new UsersPermissionData();
   }
}

export class UsersSelectReq {
  id: number = 0;
  mobile: string = '';
  organisationid: number = 0;
}
// ...
export class UsersPermissionData {
  // New editandview privileges (single boolean for each)
  editandviewDashboard: boolean = false;
  editandviewAppointments: boolean = false;
  editandviewEvents: boolean = false;
  editandviewCreateService: boolean = false;
  editandviewCreateEvent: boolean = false;
  editandviewClients: boolean = false;
  editandviewBusinessHours: boolean = false;
  editandviewStaffManagement: boolean = false;
  editandviewLocationManagement: boolean = false;
  editandviewTemplates: boolean = false;
  editandviewPaymentSettings: boolean = false;
  editandviewBusinessvalues: boolean = false;
}
// ...
export class Organisationdeletereq
{
    organisationid : number =0

    userid : number =0

    otp:string = ""
}